Move test assertions out of `BaseLpmNetworkTest`
