+++ paymentsheet/src/test/java/com/stripe/android/paymentelement/confirmation/lpms/CashAppLpmTest.kt
@@ -2,6 +2,7 @@ package com.stripe.android.paymentelement.confirmation.lpms
 
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.PaymentMethodCreateParams
+import com.stripe.android.paymentelement.confirmation.lpms.foundations.assertIntentConfirmed
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.ParameterizedRobolectricTestRunner
@@ -14,6 +15,7 @@ internal class CashAppLpmTest(
     fun `Confirm Cash App LPM`() = test(
         testType = testType,
         createParams = PaymentMethodCreateParams.createCashAppPay(),
+        assertion = ::assertIntentConfirmed
     )
 
     companion object {
